Create estudiante
This is an example on how to use class inheritance. Here, subclass Estudiante inherits atributes and methods from class Persona. There are 02 inherited methods, 01  unique method, 04 inherited atributes and just 01 unique atribute for subclass Estudiante.  The method resumen_notas spits out an academic report on all the grades that an specific student can have. It recieves a dictionary as a parameter contanining all the courses taken by a student.